додав логування

+++ ventilation_company/database/models/calc.py
@@ -2,8 +2,10 @@
 
 from __future__ import annotations
 
-from sqlalchemy import Float, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import Float, ForeignKey, Index, Integer, Numeric, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
+
+from decimal import Decimal
 
 from ventilation_company.database.base import Base
 
@@ -21,7 +23,7 @@ class CalcMaterial(Base):
     grade: Mapped[str | None] = mapped_column(String, nullable=True)
     thickness: Mapped[float] = mapped_column(Float, nullable=False)
     unit: Mapped[str] = mapped_column(String, default="мм")
-    price_per_m2: Mapped[float] = mapped_column(Float, nullable=False, default=0)
+    price_per_m2: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False, default=0)
     waste_factor: Mapped[float] = mapped_column(Float, default=1.18)
     is_active: Mapped[int] = mapped_column(Integer, default=1)
 
@@ -46,11 +48,11 @@ class CalcCalculation(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     client_name: Mapped[str | None] = mapped_column(String, nullable=True)
     client_phone: Mapped[str | None] = mapped_column(String, nullable=True)
-    markup_percent: Mapped[float] = mapped_column(Float, default=30)
+    markup_percent: Mapped[float] = mapped_column(Float, default=30)  # % — float OK
     overhead_percent: Mapped[float] = mapped_column(Float, default=15)
     labor_mode: Mapped[str] = mapped_column(String, default="hour")
-    total_cost: Mapped[float] = mapped_column(Float, default=0)
-    sale_price: Mapped[float] = mapped_column(Float, default=0)
+    total_cost: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    sale_price: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
     status: Mapped[str] = mapped_column(String, default="draft")
     created_at: Mapped[str | None] = mapped_column(String, nullable=True)
 
@@ -69,11 +71,11 @@ class CalcItem(Base):
     size_params: Mapped[str] = mapped_column(Text, nullable=False)
     quantity: Mapped[int] = mapped_column(Integer, default=1)
     area_m2: Mapped[float] = mapped_column(Float, default=0)
-    material_cost: Mapped[float] = mapped_column(Float, default=0)
-    flange_cost: Mapped[float] = mapped_column(Float, default=0)
-    labor_cost: Mapped[float] = mapped_column(Float, default=0)
-    overhead_cost: Mapped[float] = mapped_column(Float, default=0)
-    total_cost: Mapped[float] = mapped_column(Float, default=0)
+    material_cost: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    flange_cost: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    labor_cost: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    overhead_cost: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    total_cost: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
     unit_price: Mapped[float] = mapped_column(Float, default=0)
     total_price: Mapped[float] = mapped_column(Float, default=0)
