+++ ventilation_company/database/models/project.py
@@ -9,9 +9,10 @@
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Float, ForeignKey, Index, Integer, String, Text, DateTime
+from sqlalchemy import Float, ForeignKey, Index, Integer, Numeric, String, Text, DateTime
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from ventilation_company.database.base import Base
@@ -42,17 +43,17 @@ class Project(Base):
     created_at: Mapped[datetime | None] = mapped_column(DateTime, nullable=True)
     updated_at: Mapped[datetime | None] = mapped_column(DateTime, nullable=True)
     status: Mapped[str] = mapped_column(String, default="draft")
-    total_area: Mapped[float] = mapped_column(Float, default=0)
+    total_area: Mapped[float] = mapped_column(Float, default=0)  # м² — фізична величина, float OK
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
 
     # Нові поля з sqlite3-версії
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
     metadata_json: Mapped[str | None] = mapped_column(Text, nullable=True)
     drawing_path: Mapped[str | None] = mapped_column(String, nullable=True)
-    customer_price: Mapped[float] = mapped_column(Float, default=0)
-    cost_price: Mapped[float] = mapped_column(Float, default=0)
-    salary_total: Mapped[float] = mapped_column(Float, default=0)
-    profit: Mapped[float] = mapped_column(Float, default=0)
+    customer_price: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    cost_price: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    salary_total: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    profit: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
     assigned_to: Mapped[int | None] = mapped_column(Integer, nullable=True)
     created_by: Mapped[int | None] = mapped_column(Integer, nullable=True)
 
@@ -88,8 +89,8 @@ class ProjectComponent(Base):
     component_name: Mapped[str | None] = mapped_column(String, nullable=True)
     quantity: Mapped[float | None] = mapped_column(Float, nullable=True)
     unit: Mapped[str | None] = mapped_column(String, nullable=True)
-    unit_price: Mapped[float | None] = mapped_column(Float, nullable=True)
-    total_price: Mapped[float | None] = mapped_column(Float, nullable=True)
+    unit_price: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
+    total_price: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
 
     project: Mapped["Project"] = relationship(back_populates="components")
 
@@ -102,8 +103,8 @@ class ProjectMaterial(Base):
     material_name: Mapped[str | None] = mapped_column(String, nullable=True)
     quantity: Mapped[float | None] = mapped_column(Float, nullable=True)
     unit: Mapped[str | None] = mapped_column(String, nullable=True)
-    unit_price: Mapped[float | None] = mapped_column(Float, nullable=True)
-    total_price: Mapped[float | None] = mapped_column(Float, nullable=True)
+    unit_price: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
+    total_price: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
 
     project: Mapped["Project"] = relationship(back_populates="materials")
 
@@ -116,7 +117,7 @@ class ProjectWork(Base):
     work_name: Mapped[str | None] = mapped_column(String, nullable=True)
     quantity: Mapped[float | None] = mapped_column(Float, nullable=True)
     unit: Mapped[str | None] = mapped_column(String, nullable=True)
-    unit_price: Mapped[float | None] = mapped_column(Float, nullable=True)
-    total_price: Mapped[float | None] = mapped_column(Float, nullable=True)
+    unit_price: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
+    total_price: Mapped[Decimal | None] = mapped_column(Numeric(12, 2), nullable=True)
 
     project: Mapped["Project"] = relationship(back_populates="works")
